Polishing webview render logic
